requirements.txt
alembic==1.7.5
aniso8601==9.0.1
attrs==21.2.0
click==8.0.3
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-decouple==3.5
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
Werkzeug==2.0.2
zipp==3.6.0
